Fix wrong p_actions/p_rewards for training in meta-AC2 model

+++ baselines/common/runners.py
@@ -16,7 +16,7 @@ class AbstractEnvRunner(ABC):
         # Additions to support "Learning to reinforcement learn"
         self.p_actions = np.zeros((nenv), dtype=np.int32)
         self.p_rewards = np.zeros((nenv, 1))
-        self.timesteps = np.zeros((nenv, 1))
+        self.timesteps = np.zeros((nenv, 1), dtype=np.int32)
 
     @abstractmethod
     def run(self):
